requirements.txt
cryptography
python-dotenv
streamlit
pymongo
mysql-connector-python
